[GST-1203] Adds whitelist groups option
The idea is to group users together by team and role (eg. admin,
smartshop_devs, storelocatore_devs) and then use this whitelist group
option to allow certain groups of users contextually.

For example, in the base image we may only allow the admins group to be
baked in, additional keys can be added post deployment by CI on a per
role basis, so each service role inclue this role and will specify which
groups are allowed.