[GITHUB-21] Readds .ssh dir creation
During the recent update that added the ability to install ssh keys in
an alternate directory a small was removed that created a .ssh directory
by default even if no ssh auth keys are specified. Some other roles
(like the GoCD ones) rely on the creation of this dir.

This PR adds this back in and adds a test to ensure the functionality.

Closes #21.